add about content and fix build errors
