+++ public/data/about.json
@@ -1,0 +1,30 @@
+[
+  {
+    "title": "Our Expertise",
+    "icon": "https://ninjapromo.io/wp-content/uploads/2023/02/services-smm.svg",
+    "link": "https://hypefarm.io",
+    "size": { "width": "37", "height": "37" },
+    "content": "Tailored Solutions for Web3 : HypeFarm is at the forefront of Web3 marketing, offering bespoke solutions that cater to the unique needs of blockchain projects and decentralized applications. Our services span a spectrum of key areas, including: Community Management, Token Marketing, Content Creation"
+  },
+  {
+    "title": "Our Approach",
+    "link": "https://hypefarm.io",
+    "icon": "https://ninjapromo.io/wp-content/uploads/2023/02/services-seo.svg",
+    "size": { "width": "30", "height": "30" },
+    "content": "Expertise, Creativity, and Results-Driven Strategies HypeFarm blends expertise, creativity, and a results-driven approach to deliver unparalleled value in Web3 marketing. Our team of seasoned professionals is dedicated to staying ahead of industry trends, ensuring that your brand benefits from cutting-edge strategies and innovative solutions."
+  },
+  {
+    "title": "Meaningful Engagement in the Web3 Ecosystem",
+    "icon": "https://ninjapromo.io/wp-content/uploads/2022/12/ppc.svg",
+    "link": "https://hypefarm.io",
+    "size": { "width": "37", "height": "37" },
+    "content": "We don't just market; we forge connections. HypeFarm is committed to driving meaningful engagement within the Web3 ecosystem. From community interactions to strategic token marketing, our initiatives are designed to create lasting impressions and foster genuine relationships."
+  },
+  {
+    "title": "Why Choose HypeFarm?",
+    "icon": "https://ninjapromo.io/wp-content/uploads/2023/02/services-influencer.svg",
+    "link": "https://hypefarm.io",
+    "size": { "width": "37", "height": "37" },
+    "content": " Exceptional Support and Success in Web3 Marketing Trust is the foundation of our client relationships. At HypeFarm, we offer more than just marketing services; we provide exceptional support and a commitment to your success in the Web3 landscape. Partner with us, and let's embark on a journey to thrive together in the decentralized future."
+  }
+]
