+++ src/app/company/page.jsx
@@ -1,3 +1,14 @@
-import Services from "../../sections/services";
+import About from "../../sections/about";
+import { Card } from "../../components/Card";
+import aboutData from "pub/data/about.json";
 
-export default () => <Services />;
+const AboutPage = () => {
+  return (
+    <section>
+      <About />
+      <Card services={aboutData} />
+    </section>
+  );
+};
+
+export default AboutPage;
