--- src/app/contact/page.jsx
@@ -1,3 +1,0 @@
-import Contact from "../../sections/contact";
-
-export default () => <Contact />;
