+++ src/app/contactPage/page.jsx
@@ -1,0 +1,10 @@
+import Contact from "../../sections/contact";
+
+const ContactPage = () => {
+  return (
+    <section>
+      <Contact />
+    </section>
+  );
+};
+export default ContactPage;
