+++ src/app/globals.css
@@ -15,6 +15,15 @@
     --background-end-rgb: 0, 0, 0;
   }
 }
+h1 {
+  font-size: 3.5rem;
+}
+
+@media only screen and (max-width: 400px) {
+  h1 {
+    font-size: 2rem;
+  }
+}
 /* Firefox */
 * {
   scrollbar-width: thin;
