--- src/app/home/About.jsx
@@ -1,24 +1,0 @@
-const About = () => {
-  return (
-    <section
-      id="company"
-      className="bg-primary-1000 w-full text-center align-center text-white  flex flex-col justify-center align-center px-7"
-    >
-      <div className="container">
-        <h1 className="text-2xl my-10 ">
-          <span className="text-5xl">About HypeFarm </span>
-        </h1>
-        <p className="w-1/2 mx-auto mt-2">
-          HypeFarm is a leading Web3 marketing agency specializing in blockchain
-          projects and decentralized applications. We offer tailored solutions
-          including community management, token marketing, and content creation.
-          With expertise, creativity, and results-driven strategies, we drive
-          meaningful engagement in the Web3 ecosystem. Trust us for exceptional
-          support and success in Web3 marketing. Let’s thrive together.
-        </p>
-      </div>
-    </section>
-  );
-};
-
-export default About;
