+++ src/app/home/CTA.jsx
@@ -6,17 +6,16 @@ const CTA = () => {
     <section className="cta-section">
       <div className="container">
         <h2 className="text-5xl text-center">
-          Join Hypefarm on the journey to redefine Web3 marketing.
+          Join Hypefarm on the journey to redefine Web3 marketing. Let&apos;s
+          elevate your brand, engage your audience, and cultivate success in the
+          decentralized future. Contact us today to embark on this exciting
+          venture together.
         </h2>
         {/* <h2 className="text-5xl text-center">
           Partner With a Digital Marketing Agency That Delivers Growth
         </h2> */}
-        <div>
-          Let's elevate your brand, engage your audience, and cultivate success
-          in the decentralized future. Contact us today to embark on this
-          exciting venture together.
-        </div>
-        {/* <div className="text-xl my-9 text-center">
+        <div></div>
+        <div className="text-xl my-9 text-center">
           We&apos;re a full-service digital marketing agency that helps brands
           that want to get noticed and dominate their industry. We love
           marketing. We eat, sleep and breathe it. But more importantly, we love
@@ -26,7 +25,7 @@ const CTA = () => {
           seasoned enterprise, we want to be the one that helps you chart a
           course to explosive growth. If this sounds good to you, then
           let&apos;s talk.
-        </div> */}
+        </div>
         <Link href="#contact">
           <ButtonComponent content={"Speak to our team"} filled />
         </Link>
