+++ src/app/home/index.jsx
@@ -1,9 +1,9 @@
 import Hero from "./Hero";
 import Services from "../../sections/services";
 import CTA from "./CTA";
-import Contact from "../../sections/Contact";
+import Contact from "../../sections/contact";
 import Testimonials from "./Testimonials";
-import About from "./About";
+import About from "../../sections/about";
 import FAQ from "./FAQ";
 import "./home.css";
 
@@ -13,7 +13,12 @@ const Home = () => {
       <Hero />
       <Services />
       <CTA />
-      <About />
+      <section
+        id="company"
+        className="bg-primary-1000 w-full text-center align-center text-white  flex flex-col justify-center align-center px-7"
+      >
+        <About />
+      </section>
       <Testimonials />
       <FAQ />
       <Contact />
