+++ src/app/servicePage/page.jsx
@@ -1,0 +1,10 @@
+import Services from "../../sections/services";
+
+const ServicesPage = () => {
+  return (
+    <section>
+      <Services />
+    </section>
+  );
+};
+export default ServicesPage;
