--- src/app/services/page.jsx
@@ -1,3 +1,0 @@
-import Services from "../../sections/services";
-
-export default () => <Services />;
